requirements.txt
django==4.2.22
djangorestframework
pillow
plotly
weasyprint
paho-mqtt